Recurssion Reverse Nodes
