+++ NewDataStructure/OverviewDSAbyApnaCollege/CollectionFrameWork/LinkRough.java
@@ -1,61 +1,61 @@
-package NewDataStructure.OverviewDSAbyApnaCollege.CollectionFrameWork;
-import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.List;
+// package NewDataStructure.OverviewDSAbyApnaCollege.CollectionFrameWork;
+// import java.util.Iterator;
+// import java.util.LinkedList;
+// import java.util.List;
 
-class LinkRough {
-    public static void main(String[] args) {
+// class LinkRough {
+//     public static void main(String[] args) {
         
-    }
-}
+//     }
+// }
 
 
-class Linkk{
-	public static void main(String[] args) {
+// class Linkk{
+// 	public static void main(String[] args) {
 
-		// Creating a LinkedList of Integer Type
-		LinkedList<Integer> linkedList = new LinkedList<>();
+// 		// Creating a LinkedList of Integer Type
+// 		LinkedList<Integer> linkedList = new LinkedList<>();
 
-		// Inserting some Integer values to our LinkedList
-		linkedList.add(5);
-		linkedList.add(100);
-		linkedList.add(41);
-		linkedList.add(40);
-		linkedList.add(7);
+// 		// Inserting some Integer values to our LinkedList
+// 		linkedList.add(5);
+// 		linkedList.add(100);
+// 		linkedList.add(41);
+// 		linkedList.add(40);
+// 		linkedList.add(7);
 
-		// LinkedList after insertions : [5, 100, 41, 40, 7]
+// 		// LinkedList after insertions : [5, 100, 41, 40, 7]
 
-		// Calling the function to iterate our LinkedList
-		iterateUsingIterator(linkedList);
-	}
+// 		// Calling the function to iterate our LinkedList
+// 		iterateUsingIterator(linkedList);
+// 	}
 
-	public static void iterateUsingIterator(LinkedList<Integer> linkedList){
+// 	public static void iterateUsingIterator(LinkedList<Integer> linkedList){
 
-		System.out.print("Iterating the LinkedList using Iterator : ");
-
-
-		Iterator it = linkedList.iterator();
-		while(it.hasNext()){
-			System.out.print(it.next() + " ");
-		}
-
-	}
-}
+// 		System.out.print("Iterating the LinkedList using Iterator : ");
 
 
-class LList{
-    public static void main(String[] args) {
-        List<String> li=new LinkedList<>();
+// 		Iterator it = linkedList.iterator();
+// 		while(it.hasNext()){
+// 			System.out.print(it.next() + " ");
+// 		}
 
-        li.add("10");
-        li.add("20");
-        li.add("40");
-        li.add("30");
-        li.add("12");
+// 	}
+// }
 
-        for (String str : li) {
-            System.out.print(str+" => " );
-        }
-        System.out.println("NULL");
-    }
-}
+
+// class LList{
+//     public static void main(String[] args) {
+//         List<String> li=new LinkedList<>();
+
+//         li.add("10");
+//         li.add("20");
+//         li.add("40");
+//         li.add("30");
+//         li.add("12");
+
+//         for (String str : li) {
+//             System.out.print(str+" => " );
+//         }
+//         System.out.println("NULL");
+//     }
+// }
